Move spline generation to SplineCommon, make SplineDX9.cpp identical to Spline.cpp (merge later somehow)
