+++ GPU/Common/TransformCommon.cpp
@@ -1,4 +1,3 @@
-
 // Copyright (c) 2013- PPSSPP Project.
 
 // This program is free software: you can redistribute it and/or modify
