+++ GPU/Directx9/GPU_DX9.h
@@ -87,6 +87,9 @@ public:
 	void Execute_Generic(u32 op, u32 diff);
 	void Execute_Vaddr(u32 op, u32 diff);
 	void Execute_Iaddr(u32 op, u32 diff);
+	void Execute_Prim(u32 op, u32 diff);
+	void Execute_Bezier(u32 op, u32 diff);
+	void Execute_Spline(u32 op, u32 diff);
 	void Execute_VertexType(u32 op, u32 diff);
 	void Execute_VertexTypeSkinning(u32 op, u32 diff);
 	void Execute_ViewportType(u32 op, u32 diff);
