+++ GPU/GPU.vcxproj.filters
@@ -174,6 +174,9 @@
     <ClInclude Include="Common\SoftwareTransformCommon.h">
       <Filter>Common</Filter>
     </ClInclude>
+    <ClInclude Include="Common\DrawEngineCommon.h">
+      <Filter>Common</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Math3D.cpp">
@@ -269,9 +272,6 @@
     <ClCompile Include="GeDisasm.cpp">
       <Filter>Common</Filter>
     </ClCompile>
-    <ClCompile Include="GLES\Spline.cpp">
-      <Filter>GLES</Filter>
-    </ClCompile>
     <ClCompile Include="..\ext\xbrz\xbrz.cpp">
       <Filter>Common</Filter>
     </ClCompile>
@@ -324,6 +324,15 @@
       <Filter>Common</Filter>
     </ClCompile>
     <ClCompile Include="Common\SoftwareTransformCommon.cpp">
+      <Filter>Common</Filter>
+    </ClCompile>
+    <ClCompile Include="Common\DrawEngineCommon.cpp">
+      <Filter>Common</Filter>
+    </ClCompile>
+    <ClCompile Include="GLES\Spline.cpp">
+      <Filter>GLES</Filter>
+    </ClCompile>
+    <ClCompile Include="Common\SplineCommon.cpp">
       <Filter>Common</Filter>
     </ClCompile>
   </ItemGroup>
